requirements.txt
Django>=2.2.0
djangorestframework
markdown
django-filter
drf-firebase-auth
cavoke
django-logentry-admin